Add files via upload
These files are uploaded for information --- in looking at this trying to get an idea of the chipset and firmware also - it seems the ddx3216 uses a AM386 (sc) CPU as well as the sharc DSP chip. I am trying to get an idea of how the electronics work and what data can be sent via MMC rs232 and MIDI.   It seems that the device midi functionality is probably tied to limits set in the firmware, but the mixer is actaully probably fairly powerful relatively speaking if it has an AM386 cpu and sharc processors for 32 bit operations.