Add files via upload
I am guessing it is this eprom  is where the lowest level hardware information is stored -- I suspect this works with the CMOS coder decoder to boot the device -- and access the firmware. As this is only a 512 Kbit EPROM this has me think that if the firmware itself is more than 512kbit, then this must simply be a bootloader that facilates for the firmware to be loaded probably with C machine instructions for activating the memory registy --- much like a CMOS bootloader would.